Flash notifications can now span multiple requests
Using $app->flash('message'), you can now render these messages in your
templates after using seeother, etc. This makes use of sessions, which
shouldn't be a problem, but might break something.

This also adds support for multiple flash notifications.
Part of this implementation is use of __invoke which allows use of a
simple if test from Twig templates to determine if flash messages are
available:

{% if flash %}etc{% endif %}